+++ internal/helpers/helpers.go
@@ -2,7 +2,7 @@ package helpers
 
 import "math"
 
-func RoundFloat(val float32, precision uint) float32 {
+func RoundFloat(val float64, precision uint) float64 {
 	ratio := math.Pow(10, float64(precision))
-	return float32(math.Round(float64(val)*ratio) / ratio)
+	return math.Round(val*ratio) / ratio
 }
