+++ internal/structs/withdraw.go
@@ -2,6 +2,6 @@ package structs
 
 type Withdraw struct {
 	Order       string  `json:"order"`
-	Sum         float32 `json:"sum"`
+	Sum         float64 `json:"sum"`
 	ProcessedAt string  `json:"processed_at,omitempty"`
 }
